requirements.txt
bcrypt==3.1.7
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
decorator==4.4.0
Django==2.2.4
django-cors-headers==3.1.0
idna==2.8
mysql-connector-python==8.0.17
protobuf==3.9.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
